% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wine.R
\name{fit_wine}
\alias{fit_wine}
\title{Fit the wine data}
\usage{
fit_wine(wine_data)
}
\arguments{
\item{wine_data}{A wine_data class object.}
}
\value{
A wine_fit object containing the glmnet class.
}
\description{
Fit the wine data using glmnet with the mgaussian family
}
