% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wine.R
\name{load_wine}
\alias{load_wine}
\title{Load the wine data}
\usage{
load_wine(dir_path)
}
\arguments{
\item{dir_path}{The folder containing the NMR_40wines.mat and wine_labels.csv}
}
\value{
A wine_data class with the wine data.
}
\description{
Loads the wine data given a directory location.  After which, the data
colnames and rownames are assigned based on the wine_labels.  Both the
responses (Y) and features (X) are normalized by their column wise mean and
standard deviation.
}
