% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wine_data.R
\name{plot.wine_data}
\alias{plot.wine_data}
\title{Plot wine data}
\usage{
\method{plot}{wine_data}(wine_data, plot_type = NULL, ...)
}
\arguments{
\item{wine_data}{A wine_data class object.}

\item{plot_type}{A string choice of the plot type.  Options include: map, dendrogram, correlation}
}
\value{
A plot of the wine_data class.
}
\description{
Plots the wine data based on the input parameters
}
