% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wine_predict.R
\name{plot.wine_predict}
\alias{plot.wine_predict}
\title{Plot the wine prediction}
\usage{
\method{plot}{wine_predict}(pred_wine, resid_plot = F, ...)
}
\arguments{
\item{pred_wine}{A wine_predict class.}

\item{resid_plot}{A boolean to plot the residuals vs fitted or the fitted vs the observed.}
}
\value{
A plot of the wine_predict class.
}
\description{
Plots the wine prediction
}
