% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_wine_fit.R
\name{predict.wine_fit}
\alias{predict.wine_fit}
\title{Predict the glmnet results}
\usage{
\method{predict}{wine_fit}(wine_fit, wine_data, response_names = NULL, ...)
}
\arguments{
\item{wine_fit}{A wine_fit class.}

\item{response_names}{The response names to predict}
}
\value{
A wine_predict class.
}
\description{
Predict the glmnet results given all attributes (response_names = NULL) or specific results.
}
